+++ CHANGELOG.md
@@ -4,5 +4,8 @@ All notable changes to the "Ternary ⇄ If-Else Converter" extension will be doc
 
 ## Added
 - v0.0.1 Ternary ⇄ If-Else Converter
+- v0.0.2 Toggle Button + Complexity Support Increase
+    - Added Toggle Button To Enable And Disable Hover View
+
 
 ## [Unreleased]